refactor(interpreter): share numeric-index read fast path
